// packages/trading-core/src/tpsl/types/generated/tpsl-handler.ts
export interface StatusResponse {
  successful: string | boolean;
  message: string | null;
}